Created new next.js 13 app without the tailwinds framework
I have done this so we do not have to use the framework that does not work well in next.js 13, here we have fine grain control of the ui using tailwinds. Always question what you are doing, give yourself the answers... the right answers. I asked my self "why are we using this framework with next.js 13" and I gave myself the correct answer of "make it from fresh"