Login page now works
User enters email and password, form is validated. If the form is valid it gets sent to the server. the server responds with an error email or password are wrong or with a json object of user details with a jwt if the credentials are correct. The return object is stored in localstorage and in state management